Allow for long id support in jackson integration

+++ web/src/main/java/eu/okaeri/platform/web/persistence/PersistencePathSerializer.java
@@ -35,7 +35,7 @@ public class PersistencePathSerializer extends StdSerializer<PersistencePath> {
     public void serialize(PersistencePath persistencePath, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException {
         String value = persistencePath.getValue();
         if (Numbers.isInteger(value)) {
-            jsonGenerator.writeNumber(Integer.parseInt(value));
+            jsonGenerator.writeNumber(Long.parseLong(value));
         } else {
             jsonGenerator.writeString(value);
         }
